Update repository for Create method with all models

+++ src/main/resources/sql/creation_database.sql
@@ -29,53 +29,53 @@ drop
 create
 	table
 		uga.Quartier(
-			id serial primary key not null,
+			id bigint primary key not null,
 			name varchar(30)
 		);
 
 create
 	table
 		uga.Maison(
-			ID serial primary key not null,
-			idquartier serial references uga.Quartier(id)
+			ID bigint primary key not null,
+			idquartier bigint references uga.Quartier(id)
 		);
 
 create
 	table
 		uga.Date(
-			id serial primary key not null,
+			id bigint primary key not null,
 			ddate date not null
 		);
 
 create
 	table
 		uga.Heure(
-			id serial primary key not null,
+			id bigint primary key not null,
 			heure time not null
 		);
 
 create
 	table
 		uga.TypeAppareil(
-			id serial primary key not null,
+			id bigint primary key not null,
 			name varchar(30)
 		);
 
 create
 	table
 		uga.Appareil(
-			id serial primary key not null,
+			id bigint primary key not null,
 			name varchar(30),
-			idTypeAppareil serial references uga.TypeAppareil(id),
-			idMaison serial references uga.Maison(id)
+			idTypeAppareil bigint references uga.TypeAppareil(id),
+			idMaison bigint references uga.Maison(id)
 		);
 
 create
 	table
 		uga.Consommation(
-			idDate serial references uga.Date(id),
-			idHeure serial references uga.Heure(id),
-			idAppareil serial references uga.Appareil(id),
+			idDate bigint references uga.Date(id),
+			idHeure bigint references uga.Heure(id),
+			idAppareil bigint references uga.Appareil(id),
 			etat integer,
 			energy_wh integer not null,
 			primary key(
